Fixed SubtitlesScene
Now it uses the same FadeComplete()-method to go to next scene.
Variable "scenename" in LevelChanger-script is made public so that the designer can change the level-name in the Inspector.
The subtitlesScene now use the right animations to prevent glitch where the other subtitle shows up for a second before going to next scene.